requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
matplotlib>=3.5.0
pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=2.0.0 